requirements.txt
Django==2.2.8
djangorestframework==3.10.3
coreapi==2.3.3
django-filter==2.2.0
django-oauth-toolkit==1.0.0
Pillow==6.2.1
pytz==2019.3
sqlparse==0.3.0
django-tables2==2.2.1
